Fix: #74 Added error translation + updated NewsletterContent + migrated newsletter request validation to simple service
